+++ tests/testthat/queries_tests.r
@@ -1,0 +1,687 @@
+library(testthat)
+
+# ============================================================================
+# Tests for query_umls - Basic Functionality
+# ============================================================================
+test_that("query_umls requires API key", {
+  expect_error(
+    query_umls("migraine", api_key = NULL),
+    "UMLS API key is required"
+  )
+})
+
+test_that("query_umls error message includes registration URL", {
+  expect_error(
+    query_umls("test", api_key = NULL),
+    "https://uts.nlm.nih.gov/uts/license"
+  )
+})
+
+test_that("query_umls handles httr package dependency", {
+  skip_if_not_installed("httr")
+  skip_on_cran()
+  
+  # Function should exist and be callable
+  expect_true(exists("query_umls"))
+})
+
+# ============================================================================
+# Tests for query_umls - Response Structure
+# ============================================================================
+test_that("query_umls returns data frame with correct structure", {
+  skip_on_cran()
+  skip_if_not_installed("httr")
+  
+  # Expected column structure
+  expected_cols <- c("cui", "term", "semantic_type", "source", "definition")
+  
+  # Mock response when no results found
+  mock_result <- data.frame(
+    cui = NA_character_,
+    term = "test_term",
+    semantic_type = "Unknown",
+    source = "UMLS",
+    definition = NA_character_,
+    stringsAsFactors = FALSE
+  )
+  
+  expect_true(all(expected_cols %in% colnames(mock_result)))
+  expect_equal(nrow(mock_result), 1)
+})
+
+test_that("query_umls handles missing UI field", {
+  # When UI field is missing, should return NA for CUI
+  mock_result <- data.frame(
+    cui = NA_character_,
+    term = "test",
+    semantic_type = "Unknown",
+    source = "UMLS",
+    definition = NA_character_,
+    stringsAsFactors = FALSE
+  )
+  
+  expect_true(is.na(mock_result$cui))
+  expect_equal(mock_result$semantic_type, "Unknown")
+})
+
+test_that("query_umls handles missing concept data", {
+  # When concept data is incomplete, should return Unknown semantic type
+  mock_result <- data.frame(
+    cui = "C0018681",
+    term = "test",
+    semantic_type = "Unknown",
+    source = "UMLS",
+    definition = NA_character_,
+    stringsAsFactors = FALSE
+  )
+  
+  expect_equal(mock_result$semantic_type, "Unknown")
+  expect_equal(mock_result$source, "UMLS")
+})
+
+test_that("query_umls handles missing semantic types", {
+  # When semantic types are missing
+  mock_result <- data.frame(
+    cui = "C0018681",
+    term = "test",
+    semantic_type = "Unknown",
+    source = "UMLS",
+    definition = NA_character_,
+    stringsAsFactors = FALSE
+  )
+  
+  expect_equal(mock_result$semantic_type, "Unknown")
+})
+
+test_that("query_umls handles multiple semantic types", {
+  # Multiple semantic types should be comma-separated
+  semantic_types <- c("Disease or Syndrome", "Sign or Symptom")
+  combined <- paste(semantic_types, collapse = ", ")
+  
+  expect_equal(combined, "Disease or Syndrome, Sign or Symptom")
+  expect_true(grepl(",", combined))
+})
+
+# ============================================================================
+# Tests for query_umls - API Authentication
+# ============================================================================
+test_that("query_umls authentication flow structure", {
+  skip_on_cran()
+  
[…]
