+++ R/text_preprocessing.R
@@ -7,6 +7,10 @@
 #' Get dictionary cache environment
 #' @return The environment containing cached dictionary data
 #' @export
+#' @examples
+#' # Get the dictionary cache environment
+#' cache_env <- get_dict_cache()
+#' print(ls(cache_env))
 get_dict_cache <- function() {
   .dict_cache_env
 }
@@ -155,6 +159,26 @@ preprocess_text <- function(text_data, text_column = "abstract",
 #'
 #' @return A data frame with extracted entities, their types, and positions.
 #' @export
+#' @examples
+#' # Create example text data
+#' text_data <- data.frame(
+#'   doc_id = 1:3,
+#'   abstract = c(
+#'     "Migraine is a neurological disorder causing severe headache and photophobia.",
+#'     "Serotonin receptors play a role in migraine pathophysiology.",
+#'     "Sumatriptan is an effective treatment for migraine attacks."
+#'   )
+#' )
+#'
+#' # Create example dictionary
+#' dictionary <- data.frame(
+#'   term = c("migraine", "headache", "photophobia", "serotonin", "sumatriptan"),
+#'   type = c("disease", "symptom", "symptom", "chemical", "drug")
+#' )
+#'
+#' # Extract entities
+#' entities <- extract_entities(text_data, dictionary = dictionary)
+#' print(entities)
 extract_entities <- function(text_data, text_column = "abstract",
                              dictionary = NULL,
                              case_sensitive = FALSE,
@@ -410,6 +434,19 @@ extract_entities <- function(text_data, text_column = "abstract",
 #'
 #' @return A data frame containing the dictionary.
 #' @export
+#' @examples
+#' # Load a disease dictionary from local source
+#' disease_dict <- load_dictionary("disease", source = "local")
+#' head(disease_dict)
+#'
+#' # Load with custom terms
+#' custom_dict <- data.frame(
+#'   term = c("migraine", "headache", "photophobia"),
+#'   type = c("disease", "symptom", "symptom"),
+#'   id = c("D001", "D002", "D003"),
+#'   source = rep("custom", 3)
+#' )
+#' print(custom_dict)
 load_dictionary <- function(dictionary_type = NULL,
                             custom_path = NULL,
                             source = c("local", "mesh", "umls"),
@@ -596,25 +633,13 @@ load_from_mesh <- function(dictionary_type, n_terms = 200, query = NULL) {
     stop("The xml2 package is required for parsing MeSH data. Install it with: install.packages('xml2')")
   }
 
-  # Build query based on dictionary type - expanded for more categories
-  mesh_query_map <- list(
-    "disease" = "disease[MeSH]",
-    "drug" = "pharmaceutical preparations[MeSH]",
-    "gene" = "genes[MeSH]",
-    "protein" = "proteins[MeSH]",
-    "chemical" = "chemicals[MeSH]",
-    "pathway" = "metabolic networks and pathways[MeSH]",
-    "anatomy" = "anatomy[MeSH]",
-    "organism" = "organisms[MeSH]",
-    "biological_process" = "biological phenomena[MeSH]",
-    "cell" = "cells[MeSH]",
-    "tissue" = "tissues[MeSH]",
-    "symptom" = "signs and symptoms[MeSH]",
-    "diagnostic_procedure" = "diagnostic techniques and procedures[MeSH]",
-    "therapeutic_procedure" = "therapeutics[MeSH]",
-    "phenotype" = "phenotype[MeSH]",
-    "molecular_function" = "molecular biology[MeSH]"
-  )
+  # Build query based on dictionary type (using static_data)
+  if (is.null(dictionary_type) || !dictionary_type %in% names(static_data$mesh_query_map)) {
+    # If dictionary_type is not specified or not in our mapping, use a general search
+    base_query <- "mesh[sb]"
[…]
