+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   Intro_to_Literature-Based_Discovery: Intro_to_Literature-Based_Discovery.html
   Text_Preprocessing: Text_Preprocessing.html
   Work_with_Discovery_Models: Work_with_Discovery_Models.html
-last_built: 2025-09-27T04:28Z
+last_built: 2025-09-27T04:38Z
 urls:
   reference: https://liu-chao.site/LBDiscover/reference
   article: https://liu-chao.site/LBDiscover/articles
