+++ .github/workflows/test-coverage.yaml
@@ -27,8 +27,14 @@ jobs:
       - name: Test coverage
         run: |
           covr::codecov(
+            type = "tests",
+            line_exclusions = list(
+              "tests/testthat/test-slow.R",
+              "tests/testthat/test-integration.R"
+            ),
             quiet = FALSE,
             clean = FALSE,
-            install_path = file.path(Sys.getenv("RUNNER_TEMP"), "package")
+            install_path = file.path(Sys.getenv("RUNNER_TEMP"), "package"),
+            token = "${{ secrets.90c513b5-cbe7-4c8c-92f2-8aeb785b8b0b }}"
           )
         shell: Rscript {0}
