+++ tests/testthat/test-abc_model_test.R
@@ -1,0 +1,788 @@
+# Test file for abc_model.R functions
+# Tests for create_comat(), is_valid_biomedical_entity(), abc_model(), and related functions
+
+library(testthat)
+
+# Helper function to create test data
+create_test_entity_data <- function() {
+  data.frame(
+    doc_id = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
+    entity = c("migraine", "headache", "serotonin",
+               "migraine", "serotonin", "receptor",
+               "headache", "receptor", "pain"),
+    entity_type = c("disease", "symptom", "chemical",
+                    "disease", "chemical", "protein",
+                    "symptom", "protein", "symptom"),
+    count = c(2, 1, 3, 1, 2, 1, 2, 1, 1),
+    stringsAsFactors = FALSE
+  )
+}
+
+# ==============================================================================
+# Tests for create_comat()
+# ==============================================================================
+
+test_that("create_comat creates matrix with correct dimensions", {
+  skip_if_not_installed("Matrix")
+
+  entity_data <- create_test_entity_data()
+  co_matrix <- create_comat(entity_data, normalize = FALSE)
+
+  expect_s4_class(co_matrix, "Matrix")
+  unique_entities <- unique(entity_data$entity)
+  expect_equal(nrow(co_matrix), length(unique_entities))
+  expect_equal(ncol(co_matrix), length(unique_entities))
+})
+
+test_that("create_comat normalizes correctly", {
+  skip_if_not_installed("Matrix")
+
+  entity_data <- create_test_entity_data()
+  co_matrix <- create_comat(entity_data, normalize = TRUE,
+                            normalization_method = "cosine")
+
+  # Normalized values should be between 0 and 1
+  expect_true(all(co_matrix >= 0))
+  expect_true(all(co_matrix <= 1))
+})
+
+test_that("create_comat handles different normalization methods", {
+  skip_if_not_installed("Matrix")
+
+  entity_data <- create_test_entity_data()
+
+  cosine <- create_comat(entity_data, normalize = TRUE,
+                         normalization_method = "cosine")
+  jaccard <- create_comat(entity_data, normalize = TRUE,
+                          normalization_method = "jaccard")
+  dice <- create_comat(entity_data, normalize = TRUE,
+                       normalization_method = "dice")
+
+  expect_s4_class(cosine, "Matrix")
+  expect_s4_class(jaccard, "Matrix")
+  expect_s4_class(dice, "Matrix")
+
+  # Methods should produce different results
+  expect_false(identical(as.matrix(cosine), as.matrix(jaccard)))
+})
+
+test_that("create_comat stores entity types as attribute", {
+  skip_if_not_installed("Matrix")
+
+  entity_data <- create_test_entity_data()
+  co_matrix <- create_comat(entity_data)
+
+  entity_types <- attr(co_matrix, "entity_types")
+  expect_false(is.null(entity_types))
+  expect_type(entity_types, "character")
+})
+
+test_that("create_comat stores metadata", {
+  skip_if_not_installed("Matrix")
+
+  entity_data <- create_test_entity_data()
+  co_matrix <- create_comat(entity_data)
+
+  metadata <- attr(co_matrix, "metadata")
+  expect_false(is.null(metadata))
+  expect_true("n_docs" %in% names(metadata))
+  expect_true("n_entities" %in% names(metadata))
+})
+
+test_that("create_comat sets diagonal to zero", {
+  skip_if_not_installed("Matrix")
+
+  entity_data <- create_test_entity_data()
+  co_matrix <- create_comat(entity_data, normalize = FALSE)
+
[…]
