+++ R/text_preprocessing.R
@@ -41,6 +41,11 @@ preprocess_text <- function(text_data, text_column = "abstract",
                             min_word_length = 3,
                             max_word_length = 50) {
 
+  # Check if text_data is NULL or empty
+  if (is.null(text_data) || nrow(text_data) == 0) {
+    stop("text_data cannot be NULL or empty")
+  }
+
   # Add ID column if not present
   if (!"doc_id" %in% colnames(text_data)) {
     text_data$doc_id <- seq_len(nrow(text_data))
@@ -626,11 +631,13 @@ load_dictionary <- function(dictionary_type = NULL,
 load_from_mesh <- function(dictionary_type, n_terms = 200, query = NULL) {
   # Check if rentrez is available
   if (!requireNamespace("rentrez", quietly = TRUE)) {
-    stop("The rentrez package is required for MeSH queries. Install it with: install.packages('rentrez')")
+    warning("The rentrez package is required for MeSH queries. Using dummy dictionary instead.")
+    return(create_dummy_dictionary(dictionary_type))
   }
 
   if (!requireNamespace("xml2", quietly = TRUE)) {
-    stop("The xml2 package is required for parsing MeSH data. Install it with: install.packages('xml2')")
+    warning("The xml2 package is required for parsing MeSH data. Using dummy dictionary instead.")
+    return(create_dummy_dictionary(dictionary_type))
   }
 
   # Build query based on dictionary type (using static_data)
@@ -639,13 +646,6 @@ load_from_mesh <- function(dictionary_type, n_terms = 200, query = NULL) {
     base_query <- "mesh[sb]"
   } else {
     base_query <- static_data$mesh_query_map[[dictionary_type]]
-  }
-
-  if (is.null(dictionary_type) || !dictionary_type %in% names(mesh_query_map)) {
-    # If dictionary_type is not specified or not in our mapping, use a general search
-    base_query <- "mesh[sb]"
-  } else {
-    base_query <- mesh_query_map[[dictionary_type]]
   }
 
   # Add user-provided query if available
@@ -713,9 +713,9 @@ load_from_mesh <- function(dictionary_type, n_terms = 200, query = NULL) {
     # For direct MeSH search, simplify the query
     if (!is.null(query)) {
       direct_query <- query
-    } else if (!is.null(dictionary_type) && dictionary_type %in% names(mesh_query_map)) {
+    } else if (!is.null(dictionary_type) && dictionary_type %in% names(static_data$mesh_query_map)) {
       # Extract keyword from MeSH query (removing the [MeSH] qualifier)
-      direct_query <- gsub("\\[MeSH\\]", "", mesh_query_map[[dictionary_type]])
+      direct_query <- gsub("\\[MeSH\\]", "", static_data$mesh_query_map[[dictionary_type]])
     } else {
       direct_query <- ""  # Empty query will return all MeSH terms
     }
@@ -805,6 +805,7 @@ load_from_mesh <- function(dictionary_type, n_terms = 200, query = NULL) {
 
   return(combined_dict)
 }
+
 #' Process MeSH XML data with improved error handling
 #'
 #' Helper function to process MeSH XML data and extract terms
@@ -1522,8 +1523,68 @@ validate_umls_key <- function(api_key, validator_api_key = NULL) {
 create_dummy_dictionary <- function(dictionary_type) {
   message("Creating dummy dictionary for ", dictionary_type)
 
-  if (dictionary_type %in% names(static_data$dummy_dictionaries)) {
-    return(static_data$dummy_dictionaries[[dictionary_type]])
[…]
