+++ tests/testthat/test-abc_model.R
@@ -1,495 +1,744 @@
-# tests/testthat/test-abc-model.R
+# Comprehensive Test file for abc_model.R functions
+# Expanded tests to cover previously uncovered code portions
+# Tests for create_comat(), is_valid_biomedical_entity(), abc_model(), and related functions
 
 library(testthat)
-library(LBDiscover)
 
-# Helper function to create mock entity data
-create_mock_entity_data <- function() {
+# Helper function to create test data
+create_test_entity_data <- function() {
   data.frame(
-    doc_id = rep(1:10, each = 5),
-    entity = rep(c("migraine", "headache", "serotonin", "CGRP", "sumatriptan"), 10),
-    entity_type = rep(c("disease", "symptom", "chemical", "protein", "drug"), 10),
-    frequency = sample(1:5, 50, replace = TRUE),
+    doc_id = c(1, 1, 1, 2, 2, 2, 3, 3, 3),
+    entity = c("migraine", "headache", "serotonin",
+               "migraine", "serotonin", "receptor",
+               "headache", "receptor", "pain"),
+    entity_type = c("disease", "symptom", "chemical",
+                    "disease", "chemical", "protein",
+                    "symptom", "protein", "symptom"),
+    count = c(2, 1, 3, 1, 2, 1, 2, 1, 1),
     stringsAsFactors = FALSE
   )
 }
 
-# Helper function to create mock co-occurrence matrix
-create_mock_cooccurrence_matrix <- function() {
-  terms <- c("migraine", "headache", "pain", "serotonin", "CGRP", "sumatriptan",
-             "topiramate", "propranolol")
-  n <- length(terms)
-
-  set.seed(123)
-  mat <- matrix(runif(n * n, 0, 1), nrow = n, ncol = n)
-  mat[lower.tri(mat)] <- t(mat)[lower.tri(mat)]
-  diag(mat) <- 1
-
-  rownames(mat) <- colnames(mat) <- terms
-
-  entity_types <- c("disease", "symptom", "symptom", "chemical", "protein",
-                    "drug", "drug", "drug")
-  names(entity_types) <- terms
-  attr(mat, "entity_types") <- entity_types
-
-  entity_freq <- rep(5, n)
-  names(entity_freq) <- terms
-  attr(mat, "entity_freq") <- entity_freq
-
-  attr(mat, "metadata") <- list(
-    n_docs = 10,
-    n_entities = n,
-    has_types = TRUE,
-    normalization = "cosine"
+# Helper function to create test data with temporal information
+create_temporal_entity_data <- function() {
+  data.frame(
+    doc_id = c(1, 1, 2, 2, 3, 3, 4, 4, 5, 5),
+    entity = c("aspirin", "headache", "aspirin", "pain",
+               "ibuprofen", "headache", "ibuprofen", "pain",
+               "aspirin", "ibuprofen"),
+    entity_type = c("drug", "symptom", "drug", "symptom",
+                    "drug", "symptom", "drug", "symptom",
+                    "drug", "drug"),
+    year = c(2010, 2010, 2011, 2011, 2012, 2012, 2013, 2013, 2014, 2014),
+    stringsAsFactors = FALSE
   )
-
-  return(mat)
 }
 
-# Tests for create_comat
-test_that("create_comat creates matrix with valid input", {
-  entity_data <- create_mock_entity_data()
+# ==============================================================================
+# Tests for create_comat()
+# ==============================================================================
 
-  # Mock the Matrix package if needed
+test_that("create_comat creates matrix with correct dimensions", {
   skip_if_not_installed("Matrix")
 
-  result <- create_comat(
-    entity_data,
-    doc_id_col = "doc_id",
-    entity_col = "entity",
-    type_col = "entity_type",
-    normalize = FALSE
-  )
[…]
