+++ tests/testthat/test-test_queries.R
@@ -1,712 +1,242 @@
-# Comprehensive Test file for queries.R
-# This file provides complete coverage for query_umls(), query_mesh(), and enhance_abc_kb()
-# Including all previously uncovered code paths
-#
-# MOCK STRATEGY:
-# - All mocks use cycle = TRUE for safety and reliability
-# - This prevents "too many calls to mock object" errors
-# - Each test still validates the correct behavior through assertions
-# - The cycle behavior doesn't affect test validity since we verify outputs, not call counts
+# Test file for queries.R functions
+# This file tests query_umls(), query_mesh(), and enhance_abc_kb()
 
 library(testthat)
-library(mockery)
 
-# ==============================================================================
-# Helper Functions and Mocks
-# ==============================================================================
-
-# Mock HTTP responses for UMLS API
-mock_umls_auth_success <- function() {
-  mock_response <- list(
-    status_code = 201,
-    headers = list(location = "https://utslogin.nlm.nih.gov/cas/v1/tickets/TGT-123-test")
-  )
-  class(mock_response) <- "response"
-  mock_response
-}
-
-mock_umls_auth_failure <- function() {
-  mock_response <- list(
-    status_code = 401,
-    content = charToRaw("Authentication failed")
-  )
-  class(mock_response) <- "response"
-  mock_response
-}
-
-mock_umls_service_ticket <- function() {
-  "ST-123-test-ticket"
-}
-
-mock_umls_search_no_results <- function() {
-  list(
-    result = list(
-      results = list()
-    )
-  )
-}
-
-mock_umls_search_null_result <- function() {
-  list(result = NULL)
-}
-
-mock_umls_search_success <- function() {
-  list(
-    result = list(
-      results = list(
-        list(
-          ui = "C0149931",
-          name = "Migraine Disorders"
-        )
-      )
-    )
-  )
-}
-
-mock_umls_search_no_ui <- function() {
-  list(
-    result = list(
-      results = list(
-        list(
-          name = "Test Term"
-        )
-      )
-    )
-  )
-}
-
-mock_umls_concept_success <- function() {
-  list(
-    result = list(
-      name = "Migraine Disorders",
-      ui = "C0149931"
-    )
-  )
-}
-
-mock_umls_concept_null <- function() {
-  list(result = NULL)
-}
-
-mock_umls_concept_no_name <- function() {
-  list(
-    result = list(
-      ui = "C0149931"
-    )
-  )
-}
-
-mock_umls_semantics_success <- function() {
-  list(
-    result = list(
-      list(name = "Disease or Syndrome"),
-      list(name = "Clinical Finding")
-    )
-  )
-}
-
-mock_umls_semantics_empty <- function() {
-  list(result = list())
-}
-
-mock_umls_semantics_null <- function() {
-  list(result = NULL)
-}
-
-mock_umls_definitions_success <- function() {
-  list(
-    result = list(
-      list(value = "A class of disabling primary headache disorders")
-    )
-  )
-}
-
-mock_umls_definitions_empty <- function() {
-  list(result = list())
-}
-
-mock_umls_definitions_null <- function() {
-  list(result = NULL)
-}
-
-mock_umls_definitions_no_value <- function() {
-  list(
-    result = list(
-      list(source = "MSH")
-    )
-  )
-}
-
-# Mock MeSH record with all fields
-mock_mesh_record_complete <- function() {
[…]
