+++ tests/testthat/test-pubmed_search.R
@@ -54,80 +54,6 @@ create_mock_xml_response <- function() {
 </PubmedArticleSet>'
 }
 
-# New helper: Create XML with various author formats
-create_mock_xml_with_author_variations <- function() {
-  '<?xml version="1.0" ?>
-<PubmedArticleSet>
-<PubmedArticle>
-  <MedlineCitation Owner="NLM" Status="MEDLINE">
-    <PMID Version="1">11111111</PMID>
-    <Article PubModel="Print">
-      <Journal>
-        <ISOAbbreviation>Test J Abbrev</ISOAbbreviation>
-        <JournalIssue CitedMedium="Print">
-          <PubDate>
-            <MedlineDate>2023 Jan-Feb</MedlineDate>
-          </PubDate>
-        </JournalIssue>
-      </Journal>
-      <ArticleTitle>Test Article</ArticleTitle>
-      <Abstract>
-        <AbstractText>Simple abstract without labels.</AbstractText>
-      </Abstract>
-      <AuthorList CompleteYN="Y">
-        <Author ValidYN="Y">
-          <LastName>OnlyLast</LastName>
-        </Author>
-        <Author ValidYN="Y">
-          <LastName>WithInitials</LastName>
-          <Initials>AB</Initials>
-        </Author>
-        <Author ValidYN="Y">
-          <CollectiveName>Research Consortium</CollectiveName>
-        </Author>
-      </AuthorList>
-    </Article>
-  </MedlineCitation>
-  <PubmedData>
-    <ArticleIdList>
-      <ArticleId IdType="pubmed">11111111</ArticleId>
-    </ArticleIdList>
-  </PubmedData>
-</PubmedArticle>
-</PubmedArticleSet>'
-}
-
-# New helper: Create XML with alternate date/journal formats
-create_mock_xml_alternate_formats <- function() {
-  '<?xml version="1.0" ?>
-<PubmedArticleSet>
-<PubmedArticle>
-  <MedlineCitation Owner="NLM" Status="MEDLINE">
-    <PMID Version="1">22222222</PMID>
-    <Article PubModel="Electronic">
-      <Journal>
-        <ISOAbbreviation>Alt J</ISOAbbreviation>
-        <JournalIssue CitedMedium="Internet">
-          <PubDate>
-            <MedlineDate>2023 Spring</MedlineDate>
-          </PubDate>
-        </JournalIssue>
-      </Journal>
-      <ArticleTitle>Alternative Format Article</ArticleTitle>
-      <ArticleDate DateType="Electronic">
-        <Year>2023</Year>
-      </ArticleDate>
-    </Article>
-  </MedlineCitation>
-  <PubmedData>
-    <ArticleIdList>
-      <ArticleId IdType="pubmed">22222222</ArticleId>
-    </ArticleIdList>
-  </PubmedData>
-</PubmedArticle>
-</PubmedArticleSet>'
-}
-
 create_mock_search_result <- function() {
   list(
     count = 1,
@@ -141,10 +67,14 @@ create_mock_search_result <- function() {
 
 # Helper function to check if we can run integration tests
 can_run_integration_tests <- function() {
+  # Check if rentrez is available and we have internet connection
   if (!requireNamespace("rentrez", quietly = TRUE)) {
     return(FALSE)
   }
+
+  # Try a simple API call to check connectivity
   tryCatch({
+    # This is a minimal test - just try to load the rentrez package
     library(rentrez, quietly = TRUE)
     return(TRUE)
   }, error = function(e) {
@@ -169,77 +99,6 @@ test_that("parse_pubmed_xml parses XML correctly", {
   expect_equal(result$doi[1], "10.1234/test.2023.001")
 })
[…]
