+++ tests/testthat/test-additional_abc_tests.r
@@ -1,0 +1,724 @@
+# Additional tests for uncovered portions of abc_model.R
+# These tests target specific code paths not covered by existing tests
+
+library(testthat)
+
+# Helper function to create test co-occurrence matrix
+create_test_comat <- function() {
+  skip_if_not_installed("Matrix")
+
+  terms <- c("migraine", "headache", "serotonin", "receptor", "CGRP",
+             "malformation", "kinase", "pain")
+  n <- length(terms)
+
+  mat <- Matrix::Matrix(runif(n * n, 0, 0.5), nrow = n, ncol = n, sparse = TRUE)
+  # Make symmetric using Matrix transpose
+  mat <- (mat + Matrix::t(mat)) / 2
+  Matrix::diag(mat) <- 5  # Set diagonal to represent frequency
+
+  rownames(mat) <- colnames(mat) <- terms
+
+  entity_types <- c("disease", "symptom", "chemical", "protein", "protein",
+                    "disease", "protein", "symptom")
+  names(entity_types) <- terms
+  attr(mat, "entity_types") <- entity_types
+
+  entity_freq <- rep(5, n)
+  names(entity_freq) <- terms
+  attr(mat, "entity_freq") <- entity_freq
+
+  attr(mat, "metadata") <- list(
+    n_docs = 100,
+    n_entities = n,
+    has_types = TRUE,
+    normalization = "cosine"
+  )
+
+  return(mat)
+}
+
+# ==============================================================================
+# Tests for is_valid_biomedical_entity - special cases
+# ==============================================================================
+
+test_that("is_valid_biomedical_entity handles malformation special case", {
+  # Should return TRUE when claimed as disease (special exception)
+  expect_true(is_valid_biomedical_entity("malformation", "disease"))
+})
+
+test_that("is_valid_biomedical_entity handles receptor as protein", {
+  # Receptor should be recognized as protein
+  expect_true(is_valid_biomedical_entity("receptor", "protein"))
+  expect_true(is_valid_biomedical_entity("receptors", "protein"))
+})
+
+test_that("is_valid_biomedical_entity rejects analytical acronyms as chemicals", {
+  # These should be rejected as chemicals (they're analytical methods)
+  expect_false(is_valid_biomedical_entity("HPLC", "chemical"))
+  expect_false(is_valid_biomedical_entity("LCMS", "chemical"))
+  expect_false(is_valid_biomedical_entity("PCR", "chemical"))
+  expect_false(is_valid_biomedical_entity("ELISA", "chemical"))
+  expect_false(is_valid_biomedical_entity("MRI", "chemical"))
+})
+
+test_that("is_valid_biomedical_entity handles empty/invalid input", {
+  expect_false(is_valid_biomedical_entity("", NULL))
+  expect_false(is_valid_biomedical_entity(NA, NULL))
+  expect_false(is_valid_biomedical_entity(NULL, NULL))
+})
+
+test_that("is_valid_biomedical_entity handles very short terms", {
+  # Very short terms should be rejected unless they're valid acronyms
+  expect_false(is_valid_biomedical_entity("ab", "disease"))
+  expect_false(is_valid_biomedical_entity("x", "protein"))
+})
+
+# ==============================================================================
+# Tests for abc_model - validation methods
+# ==============================================================================
+
+test_that("abc_model handles different validation methods", {
+  skip_if_not_installed("Matrix")
+
+  co_matrix <- create_test_comat()
+
+  # Test pattern validation (default)
+  results_pattern <- abc_model(
+    co_matrix,
+    a_term = "migraine",
+    validation_method = "pattern",
+    exclude_general_terms = TRUE,
+    min_score = 0.01
+  )
+
+  expect_s3_class(results_pattern, "data.frame")
+})
+
[…]
