+++ tests/testthat/test-pubmed_search.R
@@ -54,27 +54,89 @@ create_mock_xml_response <- function() {
 </PubmedArticleSet>'
 }
 
-create_mock_search_result <- function() {
-  list(
-    count = 1,
-    ids = "12345678",
-    web_history = list(
-      WebEnv = "test_webenv",
-      QueryKey = "1"
-    )
-  )
+create_mock_xml_with_medline_date <- function() {
+  '<?xml version="1.0" ?>
+<PubmedArticleSet>
+<PubmedArticle>
+  <MedlineCitation>
+    <PMID>11111111</PMID>
+    <Article>
+      <Journal>
+        <ISOAbbreviation>Abbrev J</ISOAbbreviation>
+        <JournalIssue>
+          <PubDate>
+            <MedlineDate>2022 Spring</MedlineDate>
+          </PubDate>
+        </JournalIssue>
+      </Journal>
+      <ArticleTitle>Test with MedlineDate</ArticleTitle>
+    </Article>
+  </MedlineCitation>
+</PubmedArticle>
+</PubmedArticleSet>'
+}
+
+create_mock_xml_with_collective_author <- function() {
+  '<?xml version="1.0" ?>
+<PubmedArticleSet>
+<PubmedArticle>
+  <MedlineCitation>
+    <PMID>22222222</PMID>
+    <Article>
+      <Journal>
+        <Title>Collective Journal</Title>
+        <JournalIssue>
+          <PubDate>
+            <Year>2023</Year>
+          </PubDate>
+        </JournalIssue>
+      </Journal>
+      <ArticleTitle>Article with Collective Author</ArticleTitle>
+      <AuthorList>
+        <Author>
+          <CollectiveName>Research Consortium Group</CollectiveName>
+        </Author>
+        <Author>
+          <LastName>Smith</LastName>
+          <Initials>JS</Initials>
+        </Author>
+      </AuthorList>
+    </Article>
+  </MedlineCitation>
+</PubmedArticle>
+</PubmedArticleSet>'
+}
+
+create_mock_xml_with_article_date <- function() {
+  '<?xml version="1.0" ?>
+<PubmedArticleSet>
+<PubmedArticle>
+  <MedlineCitation>
+    <PMID>33333333</PMID>
+    <Article>
+      <Journal>
+        <Title>Date Test Journal</Title>
+        <JournalIssue>
+          <PubDate></PubDate>
+        </JournalIssue>
+      </Journal>
+      <ArticleTitle>Article with ArticleDate</ArticleTitle>
+      <ArticleDate>
+        <Year>2021</Year>
+      </ArticleDate>
+    </Article>
+  </MedlineCitation>
+</PubmedArticle>
+</PubmedArticleSet>'
 }
 
 # Helper function to check if we can run integration tests
 can_run_integration_tests <- function() {
-  # Check if rentrez is available and we have internet connection
   if (!requireNamespace("rentrez", quietly = TRUE)) {
     return(FALSE)
   }
 
-  # Try a simple API call to check connectivity
   tryCatch({
-    # This is a minimal test - just try to load the rentrez package
     library(rentrez, quietly = TRUE)
     return(TRUE)
   }, error = function(e) {
@@ -82,7 +144,7 @@ can_run_integration_tests <- function() {
   })
 }
 
-# Tests for parse_pubmed_xml function (doesn't require mocking)
+# Tests for parse_pubmed_xml function
 test_that("parse_pubmed_xml parses XML correctly", {
   skip_if_not_installed("xml2")
 
@@ -97,6 +159,49 @@ test_that("parse_pubmed_xml parses XML correctly", {
   expect_equal(result$authors[1], "Smith John, Doe Jane")
[…]
