+++ tests/testthat/test-text_preprocessing_additional.R
@@ -1,0 +1,597 @@
+# Additional tests for uncovered portions of text_preprocessing.R
+library(testthat)
+library(LBDiscover)
+
+# ============================================================================
+# Helper functions - MUST BE DEFINED FIRST
+# ============================================================================
+create_test_data <- function() {
+  data.frame(
+    doc_id = 1:3,
+    abstract = c(
+      "Migraine is a neurological disorder causing severe headache and photophobia.",
+      "Serotonin receptors play a role in migraine pathophysiology.",
+      "Sumatriptan is an effective treatment for migraine attacks."
+    ),
+    title = c("Migraine Study 1", "Serotonin Research", "Drug Treatment"),
+    stringsAsFactors = FALSE
+  )
+}
+
+create_test_dictionary <- function() {
+  data.frame(
+    term = c("migraine", "headache", "photophobia", "serotonin", "sumatriptan", "receptor"),
+    type = c("disease", "symptom", "symptom", "chemical", "drug", "protein"),
+    id = paste0("TEST_", 1:6),
+    source = rep("test", 6),
+    stringsAsFactors = FALSE
+  )
+}
+
+# ============================================================================
+# Test: load_from_umls - semantic type extraction edge cases
+# ============================================================================
+test_that("load_from_umls handles semantic types as nested list elements", {
+  skip_if_not_installed("httr")
+  skip_if_not_installed("jsonlite")
+  skip_on_cran()
+
+  # This tests the complex semantic type extraction logic
+  # We can't easily mock the UMLS API, so we test with invalid key
+  # to trigger error paths
+
+  result <- suppressWarnings(
+    LBDiscover:::load_from_umls(
+      dictionary_type = "disease",
+      api_key = "invalid_key_for_testing",
+      n_terms = 5,
+      semantic_types = c("T047")
+    )
+  )
+
+  expect_s3_class(result, "data.frame")
+})
+
+test_that("load_from_umls handles semantic types as data frame with uri column", {
+  skip_if_not_installed("httr")
+  skip_if_not_installed("jsonlite")
+  skip_on_cran()
+
+  # Test with multiple semantic type filters
+  result <- suppressWarnings(
+    LBDiscover:::load_from_umls(
+      dictionary_type = "drug",
+      api_key = "invalid_test_key",
+      n_terms = 5,
+      semantic_types = c("T116", "T121")
+    )
+  )
+
+  expect_s3_class(result, "data.frame")
+})
+
+test_that("load_from_umls handles semantic types with semanticType column", {
+  skip_if_not_installed("httr")
+  skip_if_not_installed("jsonlite")
+  skip_on_cran()
+
+  result <- suppressWarnings(
+    LBDiscover:::load_from_umls(
+      dictionary_type = "gene",
+      api_key = "test_key",
+      n_terms = 5,
+      semantic_types = c("T028")
+    )
+  )
+
+  expect_s3_class(result, "data.frame")
+})
+
+test_that("load_from_umls handles character semantic types in result list", {
+  skip_if_not_installed("httr")
+  skip_if_not_installed("jsonlite")
+  skip_on_cran()
+
+  # Test error handling in semantic type extraction
+  result <- suppressWarnings(
+    LBDiscover:::load_from_umls(
+      dictionary_type = "protein",
+      api_key = "invalid_key",
+      n_terms = 5,
+      semantic_types = c("T116", "T123")
+    )
+  )
+
+  expect_s3_class(result, "data.frame")
+})
+
+test_that("load_from_umls handles multiple matching semantic types", {
+  skip_if_not_installed("httr")
[…]
