+++ tests/testthat/test-abc_mode_more.R
@@ -1,0 +1,685 @@
+# Additional tests for uncovered portions of abc_model.R
+# Targeting specific code paths not covered by existing tests
+
+library(testthat)
+
+# Helper function to create test co-occurrence matrix
+create_test_matrix <- function() {
+  skip_if_not_installed("Matrix")
+
+  terms <- c("migraine", "headache", "serotonin", "receptor", "CGRP",
+             "malformation", "kinase", "pain")
+  n <- length(terms)
+
+  mat <- Matrix::Matrix(runif(n * n, 0, 0.5), nrow = n, ncol = n, sparse = TRUE)
+  mat <- (mat + Matrix::t(mat)) / 2
+  Matrix::diag(mat) <- 5
+
+  rownames(mat) <- colnames(mat) <- terms
+
+  entity_types <- c("disease", "symptom", "chemical", "protein", "protein",
+                    "disease", "protein", "symptom")
+  names(entity_types) <- terms
+  attr(mat, "entity_types") <- entity_types
+
+  entity_freq <- rep(5, n)
+  names(entity_freq) <- terms
+  attr(mat, "entity_freq") <- entity_freq
+
+  attr(mat, "metadata") <- list(
+    n_docs = 100,
+    n_entities = n,
+    has_types = TRUE,
+    normalization = "cosine"
+  )
+
+  return(mat)
+}
+
+# ==============================================================================
+# Tests for never_biomedical list and special_exceptions
+# ==============================================================================
+
+test_that("has_general_biomedical_characteristics handles never_biomedical list", {
+  # Test terms in never_biomedical that have special exceptions
+  # This should hit the special_exceptions path
+  result <- is_valid_biomedical_entity("malformation", "disease")
+  expect_true(result)
+
+  # Test terms in never_biomedical without exceptions
+  result2 <- is_valid_biomedical_entity("optimization", NULL)
+  expect_false(result2)
+})
+
+# ==============================================================================
+# Tests for similarity filtering fallback logic
+# ==============================================================================
+
+test_that("abc_model handles all B terms filtered by similarity with fallback", {
+  skip_if_not_installed("Matrix")
+
+  # Create matrix with very similar terms
+  terms <- c("migraine", "migraines", "migrainous", "migrain")
+  n <- length(terms)
+  mat <- Matrix::Matrix(0.95, nrow = n, ncol = n, sparse = TRUE)
+  Matrix::diag(mat) <- 5
+  rownames(mat) <- colnames(mat) <- terms
+
+  entity_types <- rep("disease", n)
+  names(entity_types) <- terms
+  attr(mat, "entity_types") <- entity_types
+  attr(mat, "entity_freq") <- rep(5, n)
+  names(attr(mat, "entity_freq")) <- terms
+  attr(mat, "metadata") <- list(n_docs = 10, n_entities = n, has_types = TRUE)
+
+  # This should trigger the fallback that keeps the lower half
+  # May produce warning or message about filtering
+  results <- suppressWarnings(
+    abc_model(
+      mat,
+      a_term = "migraine",
+      filter_similar_terms = TRUE,
+      similarity_threshold = 0.8,
+      min_score = 0.01
+    )
+  )
+
+  expect_s3_class(results, "data.frame")
+  # The function should still return results even if many are filtered
+})
+
+# ==============================================================================
+# Tests for suspicious entity type assignments removal
+# ==============================================================================
+
+test_that("abc_model removes suspicious entity type assignments", {
+  skip_if_not_installed("Matrix")
+
+  co_matrix <- create_test_matrix()
+
+  # Run with strict typing to trigger suspicious row detection
+  results <- suppressMessages(
+    abc_model(
[…]
