### 2025-10-03 c3dfb37
Update test file for queries
