+++ tests/testthat/test_queries_more.r
@@ -1,0 +1,463 @@
+# Additional test file for queries.R to increase code coverage
+# Focuses on uncovered portions identified in code coverage analysis
+
+library(testthat)
+
+# Mock helper functions
+create_mock_abc_results <- function() {
+  data.frame(
+    a_term = c("migraine", "migraine", "headache"),
+    b_terms = c("serotonin, CGRP", "sumatriptan", "pain"),
+    c_term = c("CGRP", "receptor", "inflammation"),
+    a_b_score = c(0.8, 0.7, 0.6),
+    b_c_score = c(0.9, 0.8, 0.7),
+    abc_score = c(0.72, 0.56, 0.42),
+    stringsAsFactors = FALSE
+  )
+}
+
+# Test query_mesh parsing logic with mocked responses
+test_that("query_mesh correctly parses mesh_id from response", {
+  skip_if_not_installed("rentrez")
+
+  # Create a mock MeSH record response - use actual newlines
+  mock_mesh_record <- paste(
+    "DescriptorUI: D008881",
+    "DescriptorName: Migraine Disorders",
+    "Tree Number: C10.228.140.546.800.525",
+    "Scope Note: A class of disabling primary headache disorders",
+    sep = "\n"
+  )
+
+  # Mock the rentrez functions
+  with_mocked_bindings(
+    entrez_search = function(...) list(count = 1, ids = "12345"),
+    entrez_fetch = function(...) mock_mesh_record,
+    .package = "rentrez",
+    {
+      result <- query_mesh("migraine")
+
+      expect_s3_class(result, "data.frame")
+      expect_equal(result$mesh_id, "D008881")
+      expect_equal(result$term, "Migraine Disorders")
+      expect_true(grepl("C10.228.140.546.800.525", result$tree_number))
+      # Verify scope_note column exists (may or may not have extracted content)
+      expect_true("scope_note" %in% colnames(result))
+      # If scope note was extracted, verify it has content
+      if (!is.na(result$scope_note)) {
+        expect_type(result$scope_note, "character")
+        expect_gt(nchar(result$scope_note), 0)
+      }
+    }
+  )
+})
+
+test_that("query_mesh handles mesh_id extraction when UI field exists", {
+  skip_if_not_installed("rentrez")
+
+  mock_mesh_record <- paste(
+    "DescriptorUI: D123456",
+    "DescriptorName: Test Term",
+    "Tree Number: A01.123.456",
+    sep = "\n"
+  )
+
+  with_mocked_bindings(
+    entrez_search = function(...) list(count = 1, ids = "123"),
+    entrez_fetch = function(...) mock_mesh_record,
+    .package = "rentrez",
+    {
+      result <- query_mesh("test")
+
+      # Test that mesh_id_match > 0 branch is executed
+      expect_equal(result$mesh_id, "D123456")
+      expect_false(is.na(result$mesh_id))
+    }
+  )
+})
+
+test_that("query_mesh extracts term name correctly", {
+  skip_if_not_installed("rentrez")
+
+  mock_mesh_record <- paste(
+    "DescriptorUI: D987654",
+    "DescriptorName: Complex Medical Term Name",
+    "Tree Number: B02.456",
+    sep = "\n"
+  )
+
+  with_mocked_bindings(
+    entrez_search = function(...) list(count = 1, ids = "987"),
+    entrez_fetch = function(...) mock_mesh_record,
+    .package = "rentrez",
+    {
+      result <- query_mesh("test")
+
+      # Test term extraction when term_match > 0
+      expect_equal(result$term, "Complex Medical Term Name")
+      expect_false(result$term == "test")  # Should be replaced with extracted term
+    }
+  )
+})
+
[…]
