+++ tests/testthat/test-text_preprocessing_more_added.R
@@ -1,0 +1,844 @@
+# Additional comprehensive tests for uncovered portions of text_preprocessing.R
+library(testthat)
+library(LBDiscover)
+
+# ============================================================================
+# Helper functions
+# ============================================================================
+create_test_data <- function() {
+  data.frame(
+    doc_id = 1:3,
+    abstract = c(
+      "Migraine is a neurological disorder causing severe headache and photophobia.",
+      "Serotonin receptors play a role in migraine pathophysiology.",
+      "Sumatriptan is an effective treatment for migraine attacks."
+    ),
+    title = c("Migraine Study 1", "Serotonin Research", "Drug Treatment"),
+    stringsAsFactors = FALSE
+  )
+}
+
+create_test_dictionary <- function() {
+  data.frame(
+    term = c("migraine", "headache", "photophobia", "serotonin", "sumatriptan", "receptor"),
+    type = c("disease", "symptom", "symptom", "chemical", "drug", "protein"),
+    id = paste0("TEST_", 1:6),
+    source = rep("test", 6),
+    stringsAsFactors = FALSE
+  )
+}
+
+# ============================================================================
+# Test: sanitize_dictionary - Complete coverage of validation logic
+# ============================================================================
+test_that("sanitize_dictionary handles all validation paths for proteins", {
+  protein_dict <- data.frame(
+    term = c("receptor", "receptors", "enzyme kinase", "short", "albumin protein"),
+    type = rep("protein", 5),
+    id = paste0("PROT_", 1:5),
+    source = rep("test", 5),
+    stringsAsFactors = FALSE
+  )
+
+  result <- sanitize_dictionary(protein_dict, validate_types = TRUE, verbose = FALSE)
+
+  expect_s3_class(result, "data.frame")
+  # "receptor" and "receptors" should be preserved as special cases
+  if (nrow(result) > 0) {
+    expect_true("receptor" %in% result$term || "receptors" %in% result$term ||
+                  any(grepl("kinase|albumin", result$term)))
+  }
+})
+
+test_that("sanitize_dictionary handles disease validation with migraine special case", {
+  disease_dict <- data.frame(
+    term = c("migraine", "cancer syndrome", "random123", "infection disease"),
+    type = rep("disease", 4),
+    id = paste0("DIS_", 1:4),
+    source = rep("test", 4),
+    stringsAsFactors = FALSE
+  )
+
+  result <- sanitize_dictionary(disease_dict, validate_types = TRUE, verbose = FALSE)
+
+  expect_s3_class(result, "data.frame")
+  # "migraine" should specifically be preserved
+  if (nrow(result) > 0) {
+    expect_true("migraine" %in% result$term)
+  }
+})
+
+test_that("sanitize_dictionary handles symptom validation with known symptoms", {
+  symptom_dict <- data.frame(
+    term = c("pain", "headache", "photophobia", "fatigue", "random_word"),
+    type = rep("symptom", 5),
+    id = paste0("SYMP_", 1:5),
+    source = rep("test", 5),
+    stringsAsFactors = FALSE
+  )
+
+  result <- sanitize_dictionary(symptom_dict, validate_types = TRUE, verbose = FALSE)
+
+  expect_s3_class(result, "data.frame")
+  if (nrow(result) > 0) {
+    known <- c("pain", "headache", "photophobia", "fatigue")
+    expect_true(any(known %in% result$term))
+  }
+})
+
+test_that("sanitize_dictionary handles biological_process with known processes", {
+  bioprocess_dict <- data.frame(
[…]
