### 2026-03-20 #725 @emilyriederer
Thanks @chaoliu-cl !

I wanted to follow up on one last topic. We recently put out our [first commentary](https://ropensci.org/blog/2026/02/26/ropensci-ai-policy/) on the use of AI in package development and review. As part of this we are in the process of updating the submission template to [request more information on authors' use of AI](https://ropensci.org/blog/2026/02/26/ropensci-ai-policy/#practical-application-for-authors).

Could you please review the discussion there and comment on the following? If AI was used, an excellent recent example of usage is found https://github.com/ropensci/software-review/issues/752#issuecomment-3938097125.

```
## Use of Generative AI

- [ ] Generative AI tools were used to produce some of the material in this submission.

If so, please describe:
If your repository includes an "agents.md" file or equivalent, please provide a link, and describe how this has been used in the development of your package
```
