### 2026-03-20 #725 @emilyriederer
Thanks, @chaoliu-cl ! Could you please provide any more details as per the linked example? For example, we are interested in which AIs were used, what sort of directions were provided (if an agents.md, skills.md, etc. exists), and what controls you applied to ensure the quality of AI-generated code.
