### 2026-03-25 #725 @emilyriederer
Thanks @chaoliu-cl ! Just to go one layer deeper, was AI used to:

- ideate / plan how to change the code
- directly edit the code
- generate tests for the code

?

Again, there is no right or wrong here! We just are trying to get a better grasp on how authors work and how this should impact of our review.
