### 2026-03-26 #725 @chaoliu-cl
Hi @emilyriederer , please see my respoonses below:

- Ideation / planning: AI was occasionally used to suggest alternative implementations for improving efficiency (e.g., vectorization strategies or refactoring specific functions), but the overall design and workflow of the package were determined by me.
- Direct code editing: Yes, AI was used to propose revisions to portions of the code to improve speed and streamline parts of the workflow (particularly in sections related to processing and linking scientific literature). All AI-generated suggestions were manually reviewed, adapted, and integrated selectively.
- Test generation: AI was used to generate tests but the actual testing and validation were reviewed by me and conducted through RStudio and Github Action.
