### 2026-05-03 #725 @maurolepore
I see @robitalec was just assigned as your editor. Looking forward to seeing this move through review.

Thanks for sharing your work with the rOpenSci community.
